form field validation

+++ app.rb
@@ -66,10 +66,12 @@ delete '/:id/return' do
   @patron = Patron.find(params.fetch("id").to_i)
   @patron_id = params.fetch("id").to_i
   @results = []
-  @return_selected = params.fetch("return")
-  @return_selected.each do |selected|
-    @book_id = selected.to_i
-    DB.exec("DELETE FROM checkouts WHERE book_id = #{@book_id};")
+  @return_selected = params.fetch("return", nil)
+  if (@return_selected != nil)
+    @return_selected.each do |selected|
+      @book_id = selected.to_i
+      DB.exec("DELETE FROM checkouts WHERE book_id = #{@book_id};")
+    end
   end
   @checked_out_books = DB.exec("SELECT book_id FROM checkouts WHERE patron_id = #{params.fetch("id").to_i};")
   erb :patron_interface
@@ -78,7 +80,11 @@ end
 patch '/:id/edit_patron' do
   @patron = Patron.find(params.fetch("id").to_i)
   name = params.fetch('name')
-  @patron.update({:name => name})
+  if (name.split('').any?)
+    @patron.update({:name => name})
+  else
+    @patron.update({:name => "#{@patron.name}"})
+  end
   @results = []
   @checked_out_books = DB.exec("SELECT book_id FROM checkouts WHERE patron_id = #{params.fetch("id").to_i};")
   erb :patron_interface
